Rename Sidebars class and methods to Widget Areas
