+++ inc/Image_Sizes/Component.php
@@ -51,7 +51,7 @@ class Component implements Component_Interface {
 			$sizes = '100vw';
 		}
 
-		if ( wp_rig()->is_primary_sidebar_active() ) {
+		if ( wp_rig()->is_widget_area_active( 'sidebar-1' ) ) {
 			$sizes = '(min-width: 960px) 75vw, 100vw';
 		}
 
@@ -85,7 +85,7 @@ class Component implements Component_Interface {
 	public function filter_post_thumbnail_sizes_attr( array $attr, WP_Post $attachment, $size ) : array {
 		$attr['sizes'] = '100vw';
 
-		if ( wp_rig()->is_primary_sidebar_active() ) {
+		if ( wp_rig()->is_widget_area_active( 'sidebar-1' ) ) {
 			$attr['sizes'] = '(min-width: 960px) 75vw, 100vw';
 		}
