+++ sidebar.php
@@ -9,7 +9,7 @@
 
 namespace WP_Rig\WP_Rig;
 
-if ( ! wp_rig()->is_primary_sidebar_active() ) {
+if ( ! wp_rig()->is_widget_area_active( 'sidebar-1' ) ) {
 	return;
 }
 
@@ -17,5 +17,5 @@ wp_rig()->print_styles( 'wp-rig-sidebar', 'wp-rig-widgets' );
 
 ?>
 <aside id="secondary" class="primary-sidebar widget-area">
-	<?php wp_rig()->display_primary_sidebar(); ?>
+	<?php wp_rig()->display_widget_area( 'sidebar-1' ); ?>
 </aside><!-- #secondary -->
